requirements.txt
channels[daphne]==4.0.0
Django==4.2.10
django-request-logging==0.7.5
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
django-ses==3.5.2
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
